#302: rework archi => create subcontainers and subcomponents + change style + move files
